Time: 0 ms (100.00%), Space: 5.9 MB (30.31%) - LeetHub

+++ 7-reverse-integer/7-reverse-integer.cpp
@@ -5,7 +5,7 @@ public:
         bool negative=x<0?1:0;
         
         int Y=abs(x);
-        long long invert=0;
+        long invert=0;
         while(Y>0){
             invert=invert*10+Y%10;
             Y/=10;
